remove redundant cast

+++ src/cvector.h
@@ -469,7 +469,7 @@ void cvector_resize(cvector *vec, size_t count, void *value)
     {
         if (vec)
         {
-            size_t cv_sz_count__ = (size_t)(count);
+            size_t cv_sz_count__ = count;
             size_t cv_sz__ = cvector_vec_to_base(vec)->size;
             if (cv_sz_count__ > cv_sz__)
             {
